requirements.txt
Flask
tensorflow
pillow
numpy
matplotlib
gunicorn
